Revert "Revert "Works with x64 & release""
This reverts commit fd4175b905e710185cc61a8deb3a3743bbd1a4e6.

+++ RemoteRendering/RemoteEncoder.cpp
@@ -361,7 +361,7 @@ void RemoteEncoder::createCuda()
 	//Setup Frameprops & Memory...
 	unsigned int widthInBytes = m_EncoderParams->iInputSize[0];
 	unsigned int height = (unsigned int)(m_EncoderParams->iInputSize[1] * 12) >> 3;
-	unsigned int cuPitch;
+	size_t cuPitch;
 	cuRes = cuMemAllocPitch(&dptr, &cuPitch, widthInBytes, height, 16);
 	handleCudaError(cuRes, "MemAllocPitch:");
